refactor : atom과 selector 파일 구조 변경 및 경로 수정
